+++ DZ8/src/user-summary.ts
@@ -1,0 +1,17 @@
+import { User } from './abstraction';
+
+// Клас для трансформації отриманих даних
+
+class UserSummary {
+    public fullName: string;
+    public companyName: string;
+    public city: string;
+
+    public constructor(user: User) {
+        this.fullName = `${user.name} (${user.username})`;
+        this.companyName = user.company.name;
+        this.city = user.address.city;
+    }
+}
+
+export { UserSummary };
